update sub account data struct and start updating controller
