+++ contracts/warp-controller/src/lib.rs
@@ -6,6 +6,8 @@ pub use crate::error::ContractError;
 
 mod execute;
 mod query;
+mod reply;
+
 #[cfg(test)]
 mod tests;
 mod util;
